requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
supabase>=2.10.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
openai>=1.50.0
numpy>=2.0.0
python-multipart>=0.0.12
google-genai>=1.0.0
